requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
azure-cognitiveservices-speech==1.33.0
sounddevice==0.4.6
numpy>=1.24.3
scipy>=1.11.3
pandas==2.1.3
openpyxl==3.1.2
python-dotenv==1.0.0
pydantic==2.6.1
pydub==0.25.1
soundfile==0.12.1 